requirements.txt
Flask==3.1.3
spacy==3.8.11
pymupdf==1.25.3
python-docx==1.2.0
sentence-transformers==5.3.0
torch==2.6.0
scikit-learn==1.8.0
pandas==2.2.3
numpy==2.2.3